Move remaining check methods in RestEntitlementsCheckAction
All future check methods should be found reflectively with the
EntitlementTest annotation.
